Refactor into separate routers
