+++ manifest.yml
@@ -1,13 +1,14 @@
 ---
 applications:
 - name: beckley-loader
-  no-route: true
-  memory: 512M
+  domain: 18f.gov
+  memory: 2G
   env:
     BECKLEY_LOADER: true
   services:
    - beckley-es
 - name: beckley
+  domain: 18f.gov
   memory: 512M
   services:
    - beckley-es
